requirements.txt
PyMuPDF==1.22.3
Pillow==9.2.0
pytesseract==0.3.10
tkinter